requirements.txt
flask==3.0.3
flask-limiter==3.8.0
openai==1.54.4
pydantic==2.9.2
python-dotenv==1.0.1
cryptography==43.0.3
pytest==8.3.3
pytest-cov==5.0.0
